requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.0.3
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
requests==2.27.1
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
gunicorn
python-dotenv
coverage
flask-bootstrap